backend/requirements.txt
argon2-cffi
django>=4.0
djangorestframework
djangorestframework-simplejwt
django-filter
drf-spectacular
gunicorn
pandas
Pillow
redis
celery[redis]
django-celery-beat
django-cors-headers
psycopg2-binary
flower
flake8
mypy
python-decouple
pytest
pytest-django
factory-boy